api/api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pytesseract
pyzbar
pillow
pandas
pyjwt
python-multipart
qrcode
